requirements.txt
playwright==1.41.0
python-dotenv==1.0.0
python-dateutil==2.8.2
colorlog==6.8.0
pytest==7.4.3
pytest-asyncio==0.21.1
